+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - `Innmind\Http\Headers::filter()`
 - `Innmind\Http\Message\Method::safe()` and `::idempotent()`
 - `Innmind\Http\Header\ContentType\Boundary`
+- `Innmind\Http\Content\Multipart`
 
 ## 6.1.0 - 2023-02-17